\name{beta.sup}
\alias{beta.sup}
\alias{beta.ave}
\alias{beta.exp}
\title{coefficient tables for F tests}
\description{Coefficient tables for p value computation of the supF-, aveF- and
expF-test (for internal use only)}
\source{Hansen B. (1997), Approximate Asymptotic P Values for Structural-Change
Tests, \emph{Journal of Business & Economic Statistics}, \bold{15}, 60-67. }
\keyword{data}
