\name{boundary.mefp}
\alias{boundary.mefp}
\title{boundary function for monitoring structural change}
\description{Computes boundary for an object of class \code{"mefp"}}
\usage{
boundary(x)}

\arguments{
  \item{x}{an object of class \code{"mefp"}.}
}

\value{an object of class \code{"ts"} with the same time properties as
the monitored process}

\author{Friedrich Leisch}

\seealso{\code{\link{mefp}}, \code{\link{plot.mefp}}}

\examples{
}
\keyword{dplot}
