\name{pvalue.efp}
\alias{pvalue.efp}
\alias{pvalue.Fstats}
\title{p value functions for structural change tests}
\description{p value functions (for internal use only)}
\usage{
pvalue.efp(x, type, alt, functional = "max", h = NULL, k = NULL)
pvalue.Fstats(x, type = c("supF", "aveF", "expF"), k, lambda)
}

\keyword{}
