\name{strucchange.internal}
\alias{sc.beta.sup}
\alias{sc.beta.ave}
\alias{sc.beta.exp}
\alias{sc.me}
\alias{pvalue.efp}
\alias{pvalue.Fstats}
\alias{monitorMECritvalTable}
\title{Internal strucchange objects}
\description{
  These are not to be called by the user.
}
\author{Achim Zeileis, Friedrich Leisch}
\keyword{internal}
