\name{strucchange.internal}
\alias{sc.beta.sup}
\alias{sc.beta.ave}
\alias{sc.beta.exp}
\alias{sc.me}
\alias{pvalue.efp}
\alias{pvalue.Fstats}
\alias{monitorMECritval}
\alias{monitorMECritvalData}
\alias{monitorMECritvalTable}
\alias{monitorRECritval}
\alias{monitorRECritvalData}
\alias{monitorRECritvalTable}
\alias{pargmaxV}
\title{Internal strucchange objects}
\description{
  These are not to be called by the user.
}
\keyword{internal}
