\name{USIncExp}
\alias{USIncExp}
\title{Income and Expenditures in the US}
\description{Data set containing the monthly personal income and personal
consumption expenditures (in billion US dollars) between January 1959 and
February 2001, which is seasonally adjusted at annual rates.}
\source{http://www.economagic.com/ }
\keyword{datasets}
